fix: category color
