+++ src/components/popup/PollPopup/Question.tsx
@@ -3,11 +3,12 @@ import Image from 'next/image';
 import styled from 'styled-components';
 import { useDarkMode } from 'usehooks-ts';
 
-import getCategoryColor from '@/styles/getCategoryColor';
+import Category from '@/types/Category';
+import { getCategoryColor } from '@/utils/categoryHelper';
 import { getCalAppWidth, getRatioSizePX } from '@/utils/sizeHelper';
 
 type QuestionProps = {
-  category: string;
+  category: Category;
   content: string;
 };
 
@@ -68,7 +69,7 @@ const TextContainer = styled.div`
   padding: 16px 0;
 `;
 
-const QuestionText = styled.span<{ $isDarkMode:boolean, $category: string }>`
+const QuestionText = styled.span<{ $isDarkMode:boolean, $category: Category }>`
   display: -webkit-box;
   -webkit-box-orient: vertical;
   -webkit-line-clamp: 4;
