+++ src/types/Category.ts
@@ -1,0 +1,3 @@
+type Category = 'romance' | 'friendship' | 'school' | 'speciality' | 'etc' | 'appearance' | 'personality';
+
+export default Category;
