+++ src/utils/categoryHelper.ts
@@ -1,15 +1,17 @@
 import _ from 'lodash/fp';
 import { DefaultTheme } from 'styled-components';
 
+import Category from '@/types/Category';
+
 type ColorType = 'fontColor' | 'backgroundColor';
 
 type Props = {
   theme: DefaultTheme;
   $isDarkMode: boolean;
-  $category: string;
+  $category: Category;
 };
 
-const getCategoryColor = (type:ColorType) => ({ theme, $isDarkMode, $category }:Props) => {
+export const getCategoryColor = (type:ColorType) => ({ theme, $isDarkMode, $category }:Props) => {
   if (type === 'backgroundColor') {
     if ($isDarkMode) {
       return _.get('color.background')(theme);
@@ -23,4 +25,4 @@ const getCategoryColor = (type:ColorType) => ({ theme, $isDarkMode, $category }:
   return _.get(`color.category.${$category}`)(theme);
 };
 
-export default getCategoryColor;
+export const getCategoryIcon = () => null;
